chore: reorganizing workflow operations
