+++ packages/client/hmi-client/src/workflow/ops/stratify-julia/stratify-operation.ts
